Quick fix Fix for fullscreen Windows
The original quickfix starts with a hard-coded 500x350 window size and
then applies zoom on that. The preview window then is also very small.

My one and only use-case is to preview full-screen windows running
video content, so this starts with 75% screen resolution as window size
which restores both the aspect ratio of my video content as well as the
original 75% screen-size upper limit